changed script and renamed file

+++ news-service/src/main/resources/db/migration/V24_1_1__basic_shema.sql
@@ -2,7 +2,7 @@ create schema if not exists develop;
 
 create table develop.news
 (
-    id             serial primary key,
+    id             bigint primary key,
     title          varchar(255) not null,
     content        text,
     published_date TIMESTAMP
@@ -10,14 +10,14 @@ create table develop.news
 
 create table develop.news_tags
 (
-    news_id int not null,
+    news_id bigint not null,
     tag     varchar(255),
     foreign key (news_id) references develop.news (id)
 );
 
 create table develop.file_attachment
 (
-    id        serial primary key,
+    id        bigint primary key,
     file_name varchar(255),
     source    varchar(255),
     url       varchar(255),
